// src/app/domains/Shared/models/User.ts
export interface User {
    name:       string;
    lastName:   string;
    nickname:   string;
    genre:      string;
    email:      string;
    phone:      string;
    category:   string;
    level:      string;
    address:    string;
    birth:      Date;
    occupation: string;
    club:       string;
    image1:     string;
    image2:     string;
}